rename getUsbDrive and type
