+++ src/preload.ts
@@ -14,7 +14,7 @@ import {
   channel as getBatteryInfoChannel,
 } from './ipc/get-battery-info';
 import { channel as getPrinterInfoChannel } from './ipc/get-printer-info';
-import { channel as getUsbDrivesChannel } from './ipc/get-usb-drives';
+import { channel as getUsbDriveInfoChannel } from './ipc/get-usb-drive-info';
 import { channel as logChannel } from './ipc/log';
 import {
   channel as mountUsbDriveChannel,
@@ -95,11 +95,11 @@ function makeKiosk(): KioskBrowser.Kiosk {
       )) as KioskBrowser.PrinterInfo[];
     },
 
-    async getUsbDrives(): Promise<KioskBrowser.UsbDrive[]> {
-      debug('forwarding `getUsbDrives` to main process');
+    async getUsbDriveInfo(): Promise<KioskBrowser.UsbDriveInfo[]> {
+      debug('forwarding `getUsbDriveInfo` to main process');
       return (await ipcRenderer.invoke(
-        getUsbDrivesChannel,
-      )) as KioskBrowser.UsbDrive[];
+        getUsbDriveInfoChannel,
+      )) as KioskBrowser.UsbDriveInfo[];
     },
 
     async mountUsbDrive(
